ci: Different LLVM version, c_ulong -> usize

+++ src/lib.rs
@@ -62,7 +62,7 @@ mod bindings {
         pub type c_ushort = u16;
         pub type c_uint = u32;
         pub type c_int = i32;
-        pub type c_ulong = u64;
+        pub type c_ulong = usize;
     }
     #[cfg(feature = "std")]
     mod types {
